Bugfix: `last_history_index` values cannot be mutated in rhai
This function used to return a `usize` which isn't consistent with `rhai`s overall approach
of using `i64` and converting to `usize` upon indexing.

Reviewers: 

Pull Request: https://github.com/haberdashPI/vscode-master-key/pull/156

+++ src/rust/parsing/src/expression.rs
@@ -451,15 +451,15 @@ fn last_index_of_history(
     context: rhai::NativeCallContext, // implicit argument, required to call a closure
     queue: HistoryQueue,
     f: rhai::FnPtr,
-) -> std::result::Result<usize, Box<EvalAltResult>> {
+) -> std::result::Result<i64, Box<EvalAltResult>> {
     let n = queue.borrow().len();
     let indices = 0..n;
     for i in indices.rev() {
         if f.call_within_context(&context, (i as i64,))? {
-            return Ok(i);
+            return Ok(i as i64);
         }
     }
-    return Ok(n);
+    return Ok(n as i64);
 }
 
 impl Default for Scope {
